Moved nodes to cpp files only
