+++ ros2_ws/src/control_cpp/src/CKeyboardOperation.cpp
@@ -5,9 +5,12 @@
 #include <unistd.h>
 #include <iostream>
 #include <fcntl.h>
+#include <termios.h>
 
-// Self Include
-#include "control_cpp/CKeyboardOperation.h"
+#include "rclcpp/rclcpp.hpp"
+#include "geometry_msgs/msg/twist.hpp"
+#include "control_cpp/NRobotModel.h"
+
 
 namespace
 {
@@ -25,62 +28,71 @@ namespace
 
 }
 
-CKeyboardOperation::CKeyboardOperation() : Node("CKeyboardOperation"), isRunning_(true)
+class CKeyboardOperation : public rclcpp::Node 
 {
-    publisher_ = this->create_publisher<geometry_msgs::msg::Twist>("robot_twist", 10);
-    timer_ = this->create_wall_timer(std::chrono::milliseconds(20), std::bind(&CKeyboardOperation::timerCallback, this));
-    key_object_ = std::make_shared<NRobotModel::KeysToVelocites>();
-    // Setup keyboard
-    // tcgetattr(STDIN_FILENO, &old_term_);
-    // new_term_ = old_term_;
-    // new_term_.c_lflag &= ~(ICANON, ECHO);
-    // tcsetattr(STDIN_FILENO, TCSANOW, &old_term_);
-
-    int flags = fcntl(STDIN_FILENO, F_GETFL, 0);
-    fcntl(STDIN_FILENO, F_SETFL, flags | O_NONBLOCK);
-
-    RCLCPP_INFO(this->get_logger(), "Keyboard Operation Node has been started");
-}
-
-CKeyboardOperation::~CKeyboardOperation()
-{
-    tcsetattr(STDIN_FILENO, TCSANOW, &old_term_);
-}
-
-void CKeyboardOperation::timerCallback() 
-{
-    // Add get key here
-    auto key = readKey();
-    if(key) {
-        processKey(*key);
-    }
-
-    geometry_msgs::msg::Twist msg;
-    msg.linear.x = speed_linear_;
-    msg.angular.z = speed_angular_;
-
-    publisher_->publish(msg);
-}
-
-void CKeyboardOperation::processKey(char key)
-{
-    if(key == 'q') 
+public: 
+    CKeyboardOperation() : Node("CKeyboardOperation"), isRunning_(true)
     {
-        RCLCPP_INFO(this->get_logger(), "Shutdown intiated by %s", this->get_name());
-        rclcpp::shutdown();
-        isRunning_ = false;
+        publisher_ = this->create_publisher<geometry_msgs::msg::Twist>("robot_twist", 10);
+        timer_ = this->create_wall_timer(std::chrono::milliseconds(20), std::bind(&CKeyboardOperation::timerCallback, this));
+        key_object_ = std::make_shared<NRobotModel::KeysToVelocites>();
+
+        // Setup Keyboard
+        int flags = fcntl(STDIN_FILENO, F_GETFL, 0);
+        fcntl(STDIN_FILENO, F_SETFL, flags | O_NONBLOCK);
+
+        RCLCPP_INFO(this->get_logger(), "Keyboard Operation Node has been started");
     }
-    else 
+
+
+    virtual ~CKeyboardOperation()
     {
-        std::tie(speed_linear_, speed_angular_) = key_object_->UpdateSpeeds(key);
+        tcsetattr(STDIN_FILENO, TCSANOW, &old_term_);
     }
-}
+    
+private:
+    void timerCallback() 
+    {
+        // Add get key here
+        auto key = readKey();
+        if(key) {
+            processKey(*key);
+        }
 
-std::optional<char> CKeyboardOperation::readKey()
-{
-    return getch();
-}
+        geometry_msgs::msg::Twist msg;
+        msg.linear.x = speed_linear_;
+        msg.angular.z = speed_angular_;
 
+        publisher_->publish(msg);
+    }
+
+    void processKey(char key)
+    {
+        if(key == 'q') 
+        {
+            RCLCPP_INFO(this->get_logger(), "Shutdown intiated by %s", this->get_name());
+            rclcpp::shutdown();
+            isRunning_ = false;
+        }
+        else 
+        {
+            std::tie(speed_linear_, speed_angular_) = key_object_->UpdateSpeeds(key);
+        }
+    }
+
+    std::optional<char> readKey()
+    {
+        return getch();
+    }
+
+    rclcpp::Publisher<geometry_msgs::msg::Twist>::SharedPtr publisher_;
+    rclcpp::TimerBase::SharedPtr timer_;
+    bool isRunning_;
+    std::shared_ptr<NRobotModel::KeysToVelocites> key_object_;
+    struct termios old_term_, new_term_;
+    double speed_linear_ = 0.0, speed_angular_ = 0.0;
+
+}; // class CKeyboardOperation
 
 int main(int argc, char *argv[]) 
 {
